Move check availableStorage, retrieveData, storeData functions to module handleData.js
