+++ assets/javascript/modules/handleData.js
@@ -1,0 +1,36 @@
+// Check if local storage available
+function storageAvailable(type) {
+  let storage;
+  try {
+    storage = window[type];
+    const x = '__storage_test__';
+    storage.setItem(x, x);
+    storage.removeItem(x);
+    return true;
+  } catch (e) {
+    return (
+      e instanceof DOMException && (e.code === 22 || e.code === 1014 || e.name === 'QuotaExceededError' || e.name === 'NS_ERROR_DOM_QUOTA_REACHED') && storage && storage.length !== 0);
+  }
+}
+
+let availableStorage;
+
+if (storageAvailable('localStorage')) {
+  // Yippee! We can use localStorage awesomeness
+  availableStorage = window.localStorage;
+} else {
+  // Too bad, no localStorage for us
+  availableStorage = null;
+}
+
+export function retrieveData() {
+  const data = availableStorage.getItem('books');
+  return JSON.parse(data);
+}
+
+export function storeData(booksArray) {
+  if (availableStorage) {
+    const jsonData = JSON.stringify(booksArray);
+    availableStorage.setItem('books', jsonData);
+  }
+}
